Tests/LibWeb/Text/expected/geometry/dompoint.txt

1. {"x":10,"y":20,"z":0,"w":1}
2. {"x":1,"y":2,"z":3,"w":4}
3. {"a":10,"b":20,"c":30,"d":40,"e":50,"f":60,"m11":10,"m12":20,"m13":0,"m14":0,"m21":30,"m22":40,"m23":0,"m24":0,"m31":0,"m32":0,"m33":1,"m34":0,"m41":50,"m42":60,"m43":0,"m44":1,"is2D":true,"isIdentity":false}
4. {"x":750,"y":1060,"z":0,"w":1}
5. {"x":750,"y":1060,"z":0,"w":1}
6. {"x":750,"y":1060,"z":0,"w":1}
7. Exception: TypeError
